message-processor/requirements.txt
requests
beautifulsoup4
lxml
redis
rq
pydub
ffmpeg
PyYAML
speechrecognition
vosk